Pulled out the ClusterClock interface to reduce NodeEngine dependencies
