+++ hazelcast/src/main/java/com/hazelcast/cluster/ClusterClock.java
@@ -1,0 +1,15 @@
+package com.hazelcast.cluster;
+
+public interface ClusterClock {
+
+    /**
+     * Returns the cluster-time.
+     * <p/>
+     * TODO: We need to document what cluster time really means and what is can be used for.
+     *
+     * @return the cluster-time.
+     */
+    long getClusterTime();
+
+    long getClusterTimeDiff();
+}
